refactoring: filepaths to utils
