+++ project/akishchenko/dataset_results_count.py
@@ -1,17 +1,5 @@
-import os
 from datasetutils import CsvStats
-
-def get_filepaths(directory):
-    file_paths = []  # List which will store all of the full filepaths.
-
-    # Walk the tree.
-    for root, directories, files in os.walk(directory):
-        for filename in files:
-            # Join the two strings in order to form the full filepath.
-            filepath = os.path.join(root, filename)
-            file_paths.append(filepath)  # Add it to the list.
-
-    return file_paths  # Self-explanatory.
+from datasetutils import get_filepaths
 
 if __name__ == '__main__':
     files = get_filepaths("results/")
